Refactor generateSine funtion and update comments.
