+++ src/server/data/generateTestData.js
@@ -3,22 +3,21 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 /**
- * 
- * I only need three functions to export to test.
- * I only need to export one function that does those function
- * functions do sequentially requested.
- * 	I need to incorporate the period of the sine wave within each period.
- * 	Which has the purpose of giving more control to each of those sine waves.
- * First I need a function that generates the dates from one period to another at a 
- * certain time-step. 
+ * generateTestData.js exports four functions: 
+ * generateDates,
+ * generateSine, 
+ * write_to_csv,
+ * generateSine 
+ *
  */
 
-//import
+// Imports
 const fs = require('fs');
 const stringify = require('csv-stringify');
 const _ = require('lodash');
 const moment = require('moment');
 
+// Global Constants
 var TWO_PI = Math.PI * 2;
 
 /**
@@ -120,14 +119,17 @@ function _generateSineData(startTimeStamp, endTimeStamp, options = { timeStep: {
 /** data needs to be of form
  * 
  * [
+ * ...
  * [...row]_generateSineData(startTimeStamp, endTimeStamp),
  * [...row],
  * [...row],
- * 
+ * ...
  * ]
+ * 
+ * Sources: 
+ * https://csv.js.org/stringify/api/
+ * https://stackoverflow.com/questions/2496710/writing-files-in-node-js
  */
-// https://csv.js.org/stringify/api/
-// https://stackoverflow.com/questions/2496710/writing-files-in-node-js
 function write_to_csv(data, filename = 'test.csv') {
 	stringify(data, (err, output) => {
 		if (err) {
@@ -140,7 +142,7 @@ function write_to_csv(data, filename = 'test.csv') {
 			console.log("The file was saved!");
 		});
 	});
-}
+} // write_to_csv(Array[Array[String, String]],String)
 
 /**
  * 
@@ -150,9 +152,10 @@ function write_to_csv(data, filename = 'test.csv') {
  * 
  * Creates a csv with sine data 
  */
-function generateSine(startTimeStamp, endTimeStamp, options = { timeStep: { hour: 12 }, period: { day: 1 } }) {
-	write_to_csv(_generateSineData(startTimeStamp, endTimeStamp, options));
-}
+function generateSine(startTimeStamp, endTimeStamp, options = { filename: 'test.csv', timeStep: { hour: 12 }, period_length: { day: 1 } }) {
+	const chosen_data_options = { timeStep: options.timeStep, period_length: options.period_length };
+	write_to_csv(_generateSineData(startTimeStamp, endTimeStamp, chosen_data_options), options.filename);
+} // generateSine(String,String,Object)
 
 module.exports = {
 	generateDates,
